Push saas deploy prometheus status from Tekton
Added a task that pushes to Prometheus a metric with the status of the
openshift-saas-deploy task.

Since we're still on Tekton pipelines v0.22.0, we don't have the
aggregate of all the status in $(tasks.status) but it is not needed for
the moment.

The metric name has been chosen to be generic and it reflects the fact
that we only keep the status of one of the tasks asociated to the
PipelineRun.

The prefix of the PipelineRun is kept as a label to have just one
timeseries per pipelinerun. To make it unique we have followed the
pattern in openshift_saas_deploy_trigger_base.construct_tekton_trigger_resource
so that we always get the status of the last run.

part of APPSRE-3320

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>