Final version of tekton push gateway metrics
There was a general misconception in the concepts that drove previous
implementation. The most important thing when pushing metrics through
the PushGateway is the group identified in the url. When POSTing metrics
through it, we're always going to overwrite whatever was there no matter
what the labels in the metrics were there.

Having the above into account, it makes sense to keep the pipelinerun as
it is, remove the pipelinerun_prefix approach and push into a
combination of saas_file and env that we know it has to be unique.

part of APPSRE-3320

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>