Adjust the interval used by OCM crash alerts
Currently the alert fires when a pod has been restarted two or more
times in a period of five minutes. But Kubernetes will by default
restart pods with less frequency than that, due to the exponential back
off. Net result is that the alert doesn't trigger when the pod is
crashing inmediately when it starts. This patch changes the period to
ten minutes to address that.

Related: https://issues.redhat.com/browse/SDA-2856
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>