Explicitly force type OCM secret key to string
This patch uses the `!!str` tag to force the tag of the secret key that
contains the Pendo key of the OCM clusters management service of the
integration v3 environment. The value of that key is empty in Vault and
witout this the key isn't added to the Kubernetes secret.