Make service mandatory for prometheus alert labels
We use service as our main label to decide routing. With this change we
force every alert defined in app-interface to have service defined. This
needed a schema change. Since only alert labels are used in routing, it
only make sense to impose restrictions there. Group rules can now have
two different objects:

* Recording rules: labels are just an object
* Alerts: labels are an object very much restricted

A few new labels have been added for alerts that we managed. Changes in
the AlertManager routing tree were needed in order to account for this
services. An explcit matcher to route to AppSRE's slack has been added
for the non-critical alerts (this makes the tree more understandable).

Finally this will make routing to jiralert and slack feasible, because
alerts that are caught by the catch-all route cannot be routed to two
different places.

part of APPSRE-2893

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>