Bug 1980753: Update the 4.7 rootfs url to match the iso
Upstream PR: https://github.com/openshift/assisted-service/pull/2189

This will solve this issue for the stage and prod environments as
the default ocp config is not actually broken in the version we have
deployed and the bug exists in these overrides instead.

Needs to be merged and deployed before
https://github.com/openshift/assisted-service/pull/2230 which will force
the minimal iso template to be regenerated with the new rootfs url.