Don't alert on repeated OCM cluster installation errors
Currently the cluster installation error fires when a cluster is updated
to the `error` state, regardless of what was the previous state. As the
service will repeatedly retry some failed operations this results in the
alert firing multiple times, which adds noise. This patch changes the
alert so that it only fires when the state transations to `error` from
something other than `error`.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>