Replace the redis version on quayio-stage
It turns out that terraform wants the engine version to be a specific
one, whereas starting from redis 6, AWS wants a rolling version. The
solution to this seems to be let terraform do its thing with 6.x and
when it starts complaining that there are no modifications to do, go
back to 6.0.5.

References:

https://github.com/hashicorp/terraform-provider-aws/issues/15625
https://github.com/hashicorp/terraform-provider-aws/issues/16564