Promote hive 34d8b527b80acfb55701034adfff9caf978e3326
https://github.com/openshift/hive/compare/625164b626c058be6278463ad6a2201bbc76c2a7...34d8b527b80acfb55701034adfff9caf978e3326

---
```
commit: 4d8d83dcd46406bc4ff1d6c7708eb7e07dc23727
author: Eric Fried
Actually start hibernated machines

PR #1477 introduced a regression whereby it wouldn't start (non-fake)
machines in most cases. Fix.

fixes bz #1989129

commit: afaff39ed2fba9acc9cf556ec11c634cbdf6cba0
author: Eric Fried
Make owned CRD order deterministic in CSV

Sort the list of CRDs when generating the CSV for the operator bundle,
for easier diffing from one version to the next.

commit: 28b65917986be39288621298992155dde796ee2b
author: Suhani
Hibernation bug with fake clusters

As a part of previous change to support hibernation for fake clusters,
it was found that
- some spec and/or status updates were missing
- code coverage for hibernation due to powerstate change (and not
  hibernateafter) was missing
- switching to running condition from initialized hibernation condition
  was not covered

This commit rearranges some code to resolve the above stated bugs

commit: 8df60ac1564005c882fc40ec69bd799e88ea9768
author: Eric Fried
Claim: Add ClusterRunning=>ClusterDeleting condition

This is a UX improvement to reflect that a CD is deleting in the status
of its assigned Claim: We set the ClusterRunning condition to False with
Reason: ClusterDeleting.

Note that, although this triggers on both deletion timestamp and the
`remove-claimed-cluster-from-pool` annotation, we are unlikely to notice
in the latter case because the claim itself is deleted almost
immediately in that code path. However, the code as written accounts for
a future where that deletion may wait on some other finalizer.

commit: fd6d5aba2fce5af7d7df29df4ddf0d216af7e1a9
author: Eric Fried
Tighten TestReconcileClusterClaim

We were skipping some chunks of TestReconcileClusterClaim if the claim
was deleted. However, there are still things we should be checking in
that case -- like the state of the assigned ClusterDeployment. Tweak the
test driver to support this.

commit: 5422b69102dd0a57fe4144012ce041b5cf88ec37
author: Abhinav Dahiya
remotemachineset: aws: fallback to using kubernetes.io/role/elb tag for public subnets when route tables are not enough

Currently the discovery codepath only decides whether a subnet is public
by using the routetables.

There are many network deployments which do not directly attach an igw
to the public subnet, but usage of these subnets is clearly defined by
the user by tagging these subnets with PublicSubnetTagName. So it
would be benefitial if the discovery codepath respect the indication on
the public subnets set by the admins of the cluster.

Similarly the aws-load-balancer-controller [1] now solely depends on
PublicSubnetTagName to discover the public subnets for ELB.

In this change, we update the isSubnetPublic function such that when it
fails to decide that a subnet is public using route table rules, if
checks if the subnet has the PublicSubnetTagName, and if the tag exists
and equal to empty("") and "1" value [2] it returns true. So the main
check is
still route tables based, and the tag based becomes a fallback.

This is based on a change that is also being floated in upstream [3].

Since the upstream change hasn't been merged. This only helps creating
machines in these subnets. If a user was to create a Service type
Loadbalancer, it would still require that some subnets exists which are
identifable as public with just route tables.

[1]:
https://github.com/kubernetes-sigs/aws-load-balancer-controller/blob/main/pkg/networking/subnet_resolver.go#L114
[2]:
https://docs.aws.amazon.com/eks/latest/userguide/network-load-balancing.html
[3]: https://github.com/kubernetes/kubernetes/pull/103861

```