feat: add the observatorium token refresher network policy to KAS Fleet Manager to production
This change is required to mitigate one of the threats identified by the Managed Kafka Security team around the Observatorium token refresher authentication flow. We now have a tenant on the Managed Service Tenant (MST) Observatorium which uses sso.redhat.com as authentication. As part of the migration to use this new Observatorium instance, a token refresher is needed to handle fetching and refreshing access tokens from sso.redhat.com so that the KAS Fleet Manager can send request to the MST Observatorium instance. All requests from KAS Fleet Manager to MST Observatorium will be directed to this token refresher. The token refresher then sets the token in the received request and forwards that onto MST Observatorium.

A network policy should be in place to protect access to this token refresher so that only the KAS Fleet Manager pods can communicate with it.

JIRA: https://issues.redhat.com/browse/MGDSTRM-1633