Get clusters without namespace in database
Some time ago we changed the code to store in the database the namespace
and cluster deployment name, so that we don't need to query by labels in
order to find them. We preserved the code for that kind of query by
labels in for old clusters. We would like to remove it now, and before
doing that we need to check if there are still clusters that don't have
the namespace and name in the database.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>