Disable auto-rebasing in MRs against service/managed-tenants.
We and our tenants currently have a lot of work when MRs get autorebased
because qontract-reconcile removes the approval label after a rebase
against the target branch and it happens very often.

Reading the code of gitlab-housekeeping gave me the impression that
disabling automatic rebasing will not make qontract-reconcile refrain
from merging the MR (as long as it was properly approved). The merges
will just be regular merges and not slick-looking fast-forward
merges.

Let's give it a try! 🤞

Signed-off-by: Josh Gwosdz <jgwosdz@redhat.com>