feature: remove external job for "sources-api-go"
Until now both external and internal jobs were kept being launched, as a
way to be able to compare the two and understand why the internal ones
were failing and the external ones were not.

As we've already refined our internal job to make it succeed, the
external one is no longer needed.