Update branch used by rhsm pr check for pre-build merge
I suspect, but cannot confirm, that this is currently preventing the
builds from being triggered.

We do not have a master branch, and the more appropriate branch to merge
to for us and our current workflow is actually develop. (We also have
main but only use main for release prep and tagging).

Similar to !17884