Add RetryReasons to HiveConfig (stg/int)
Hive recently [added](https://issues.redhat.com/browse/HIVE-1662)
support for limiting provision retries based on failure reasons parsed
out of the install logs. It is pointless to retry on the vast majority
of failure classes (access, quota, misconfiguration, etc.).

With this commit we contend that the only class of failure where it is
reasonable to retry is throttling (rate limiting). After this change,
any other failure will stop the provision immediately, irrespective of
InstallAttemptsLimit.

This commit changes only staging and integration clusters. The
underlying feature is planned to be promoted to production ~20211115; a
similar change can be made to production HiveConfigs after that time if
all goes well.

HIVE-1662