RetryReasons for HiveConfig
Hive recently [added](https://issues.redhat.com/browse/HIVE-1662)
support for limiting provision retries based on failure reasons parsed
out of the install logs. It is pointless to retry on the vast majority
of failure classes (access, quota, misconfiguration, etc.).

With this commit we contend that the only class of failure where it is
reasonable to retry is throttling (rate limiting). We are also
temporarily including a workaround for OSD-9020 whereby we sometimes
spuriously see invalid project ID failures in GCP.

After this change, any other failure will stop the provision
immediately, irrespective of InstallAttemptsLimit.

Note: The configuration for AWS throttling was previously added to
staging and integration via 70cd9e7a13 / !28613.

OSD-9043