Promote hive 8f0773c53b8463b73fa3171cebcfe6abf4a48107
https://github.com/openshift/hive/compare/f704abe069fcf9c69dc0c62841b21fd662032082...8f0773c53b8463b73fa3171cebcfe6abf4a48107

---
```
commit: 9cd299ebdf364f26e683686ce990ab5e345685b7
author: Eric Fried
Static name for managed DNS ConfigMap

When managed DNS is enabled, we copy the relevant data from HiveConfig
to a ConfigMap in the target namespace. Previously this ConfigMap was
dynamically named `managed-domains-XXXXX` and we would find it via a
label. For consistency with other similar ConfigMaps used by Hive, this
commit switches it to use a static name, `hive-managed-domains`,
instead. For upgrade purposes, we have to keep the old lookup mechanism
in place until we can be sure we've flushed out all instances of the old
naming scheme.

commit: 75b487e158f853df60ee48534a767e9b8670b7fa
author: Eric Fried
operator: DRY computing hashes

The operator computes hashes (MD5 sums) of ConfigMaps and Secrets and
uses those hashes to annotate the Deployments/StatefulSet for its child
controllers such that, if any of those ConfigMaps/Secrets change, the
annotations change, and the controllers get redeployed to pick up the
changes.

The code to compute these hashes, simple though it was, was duplicated
in a number of places. Consolidate to a single generic function.

commit: 5fbedbf63fdc8aa201701b2e898f7c84ebd034e1
author: Joel Diaz
clarify when the ProvisionStopped condition would be set

It is meant to indicate that at least one cluster provision was
attempted.

```