Set IMAGE_SERVICE_BASE_URL in integration
This will instruct assisted-installer to use the image service instead
of generating and uploading the iso to s3.

This will cause any new download url to point to the configured image
service rather than s3. Existing download urls (to s3) will remain valid
until they expire. At that point a new url will need to be generated.
This can be triggered using the same generate cluster iso endpoint that
is currently in use.

https://issues.redhat.com/browse/MGMT-7716
https://issues.redhat.com/browse/SDE-1497