Deploy the assisted-image-service to stage
This also creates a route specifically for the `/images` path for the
same host as the existing dns entry. This will direct traffic to the
image service for `/images`, but to assisted-service for every other path.

Additionally this adds the gateway configuration to expose the image
service on `api.stage.openshift.com`

https://issues.redhat.com/browse/MGMT-7716