Add metrics monitoring to assisted image service
This commit deploys the monitoring template for the
assisted-image-service so that we will gather metrics.

This also moves the stage version of the image service up to pull in the
required changes to the deploy template (mostly object metadata
changes). These were required by the new monitoring template.

Monitoring template added in
https://github.com/openshift/assisted-image-service/pull/48

https://issues.redhat.com/browse/MGMT-7715