Remove `ocm-envoy`
The `ocm-envoy` project was intended to build an image containing the
Envoy proxy with the minimal set of dependencies needed by OCM in order
to reduce the attack surface. But that was never completed, and the
image was never used or updated, which actually made it less secure.
This patch removes it completely. It wasn't ever used by any
OCM service, so nothing else needs to change.

Related: https://docs.google.com/document/d/1gmgViYV58O6mX8VlvHT1pADEfkKqMuv7xIiQjQBjtq8
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>