Add routing for the image serivce in prod
This route is specifically for the `/images` path for the same host as
the existing dns entry. This will direct traffic to the image service
for `/imges`, but to assisted-service for every other path.

Additionally this adds the gateway configuration to expose the image
service on `api.openshift.com`

Similar change for stage: 86b705910cc786c4b9d9e48deb286e257994836a

https://issues.redhat.com/browse/MGMT-7716