Remove OCM Kafka instances
This patch reverts commits ee33a9f0c6 and 672ef4ba15. Those commits added Kafka
instances for integration, stage and production environments of OCM because we
intended to use them for the upcoming HyperShift support. But we have decided to
release the HyperShift support without using Kafka initially. After the initial
release (which should be around Jun 2022) we will split the service into a core
component and separate provisioners, using Kafka for the communications between
them. But meanwhile we don't need to pay the cost of those instances, as we will
not be using them.

Related: https://issues.redhat.com/browse/SDE-1586
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>