Disable tekton monitoring from configuration
We were doing it by hacking the ServiceMonitor.

This is a first step in controlling Tekton configuration. We will then
tailor the monitoring to our needs.

The TektonConfig object has been copied as-is from the default one added
by the operator. Only the `enableMetrics` key has been set to "false"

part of APPSRE-4135

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>