Enable gauge metrics for Pipeline/Tasks in stage
The idea is to enable a minimum amount of metrics that could be useful
without making the cardinality of these metrics explode.  This is the
level that jenkins gave, the duration of the last run of the task. Since
we don't control the buckets size and there aren't that many runs, I
don't think keeping histograms is that useful.

part of APPSRE-4135

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>