Add github app-sre-trigger-bot
We intend to use this bot for Jenkins SCM polling from ci-ext fot the
moment. We are seeing lots of cache misses in github mirror that appear
to come from jenkins polling. These are mainly produced by trying to
assert the membership of users that don't belong to certain orgs, e.g.

https://api.github.com/orgs/openshift/members/dependabot[bot]

Since github mirror cannot cache 404s, then this cache misses always add
to the quota. Using this bot in jenkins will make sure that regular
github mirror usage works fine whereas some potential throttling in
jenkins SCM is more tolerable.

https://coreos.slack.com/archives/GGC2A0MS8/p1644510636645029

In ci-int there are a number of private repositories, we will see what
we can do with that.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>