pin aws provider version
The latest AWS provider version is 4.x, so if the specific provider
version is not specified, the intstructions will fail with this error:

```
Error: Value for unconfigurable attribute

  on terraform-setup.tf line 23, in resource "aws_s3_bucket" "terraform-bucket":
  23: resource "aws_s3_bucket" "terraform-bucket" {

Can't configure a value for "versioning": its value will be decided
automatically based on the result of applying this configuration.

Error: Value for unconfigurable attribute

  on terraform-setup.tf line 25, in resource "aws_s3_bucket" "terraform-bucket":
  25:   acl = "private"

Can't configure a value for "acl": its value will be decided automatically
based on the result of applying this configuration.
```

That is due to a breaking change in 4.x:
- https://github.com/hashicorp/terraform-provider-aws/issues/23103
- https://registry.terraform.io/providers/hashicorp/aws/latest/docs/guides/version-4-upgrade#versioning-argument

The ideal solution is to port these tf bootstrapping instructions to
4.x, but until that work is done, this needs to be fixed.