Gateway release (prod)
There was a test change (non-functional) that was added after the last stage
deploy. That's the reason for the bump in pre-prod as well. This is really
meant to serve as the production deploy though, pulling in all changes detailed
in the release doc: https://docs.google.com/document/d/1-wBQu_Xtp-hYPS5LQ4pMw1kAzMoEv1bfnjSQxR3jKFY/edit#