Add apicast-tests endpoints to Turnpike for testing
We'll start to use apicast-tests for Turnpike test automation, as we do in the
external gateway. The `/ping` endpoint will be used for general tests requiring
authN, but no authZ.

The `/headers` endpoint will be used to enforce authZ. We're waiting on a service
cert to enforce x509 authZ and will update once that's available.