[prod]rhobs: Update image for AMS remote-write proxy
The previous image has many vulnerabilities and is
built from https://github.com/observatorium/receive-proxy
which has been marked by https://github.com/observatorium/receive-proxy/issues/2
as a repo that will be removed in the not so distant future.

The underlying image was a basic nginx reverse proxy so
this change moves to use an existing quay/app-sre image
built atop ubi8.