Assigned contract version to all sre checkpoints
This is done retroactively, so no real good way of doing it. We have
been using a non-released version of the contract lately, so for all
2021 checkpoints I have used the last 2021 version and for 2022 the
first released, since the v2022.04.25 version reflects the changes made
to remove DVO checks from SRE checkpoints, which was not in place when
all the 2022 wave of checkpoints has been completed.

Part of APPSRE-4788

Signed-off-by: Rafael Porres Molina <rporres@gmail.com>