Add the special managment org ID to KFM SaaS yaml
For both production and stage environments.

JIRA: https://issues.redhat.com/browse/MGDSTRM-7969

Special Managment Org - is created in sso.redhat.com. This is used
as the OrgID for the service accounts created by the KFM service. These are
internal service accounts such as Canary, Fleetshard e.t.c. OrgIDs are requirement
for service accounts created in sso.redhat.com for it to be managed. Hence this change.