design doc service maintenance planning and announcements
AppSRE tenants want to announce maintenances for their services on status.redhat.com but have no or limited access to the management capabilities for statuspage.io.

This design doc proposes self service maintenance announcements for statuspage.io and also a process for maintenance planning so AppSRE is involved or at least aware.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>