data/services/cincinnati/cicd/ci-int/saas: Three stage replicas
We'd opened with two replicas for both production and stage in
7c1fb4ade1 (change replicas for cincinnatti to 2, 2020-09-02).  We'd
scaled this up to three for stage performance testing in 0aa3bf99a7
(Cincinnati stage: bump replicas to 3, 2022-02-22) and then rolled
back to two after testing in 46d9e225ba (Revert "Cincinnati stage:
bump replicas to 3", 2022-02-22).

However, the deployment validation operator is grumpy about two
replicas [1]:

  Alert: deployment_validation_operator_minimum_three_replicas Validation error for Cincinnati: Indicates when a deployment uses less than three replicas

and we're fine moving to three to keep the DVO happy.

[1]: https://issues.redhat.com/browse/OTA-648