Split dynamic DVO prometheus rules
The result of the original dynamic rule is too big. The Prometheus Rules
objects are converted to a configmap by the Prometheus operator, which
complains if the size of the resulting object is too big (> 512Kb).

The chosen split key (DVO check) is not perfect, as potentially rules
could grow with many new namespaces up to a point in which we go beyond
the max size. The perfect split would have been by namespace, but that
would have forced us to add many objects in one file, which is not
supported by openshift resources integration.

part of APPSRE-5640