Add new jira priority mapping for ASIC to attempt to help with alert spam
The reconcile channel is spammed every minute that there are active QontractReconcileError
alerts. This behavior seems to be controlled by repeat_interval in Alertmanager. The only
routes setting an interval of 1m are jiralert and deadman (the latter being irrelevant here).

Before this change, ASIC jiralerts were only matching the root jiralert route because there
was no match setup for a critical-fts severity. It seems that we starting seeing alert spam
shortly after the changes to dynamically generate jiralert configs, which added severity
to the mapping. This may be relevant due to this note in the Alertmanager docs:

"If an alert does not match any children of a node (no matching child nodes, or none exist),
the alert is handled based on the configuration parameters of the current node."

What I suspect could be happening is that the alert is matching the root jiralert route, but
since it doesn't match any child nodes, the alert is inheriting the config of the root jiralert
node (notably repeat_interval: 1m) and then continues on (continue: true) to match other routes,
including those that have Slack notifications receivers.

At the very least, this should fix that ASIC tickets aren't created for these alerts anymore.