chore: check the tenant ID for a particular EBS and OrgId numbers
It seems that we are using the wrong tenant ID when running some queries
on production for the Cost Management team. As the tenant ID comes from
the identity headers, I would like to run a query to see to wich tenant
ID do the EBS account number and OrgId number are tied to.

That can give us a clue about what is going on.