chore: get the applications and app auths related to duplicated tenants
The purpose of these queries is to see how many applications and
application authentications might be related to duplicated tenants, so
that we can have an idea of the difficulty for fixing this.

The previous query didn't work because apparently the quotes are
stripped from the queries when being executed, and that caused an issue
with the column aliases.