fix alerting rules for cincinnati stage
fix the alerting rule for detecting upstream scrapes errors as
the 5m time interval was unable to detect errors, bumping it to
15m.

also add a new alert to detect the staleness of update graph.
with severity `warning` we'll be informing the app-sre along
with the cincinnati team about the staleness of the graph.