store commit sha based app-interface bundles in s3
additionally to the `data.json` named key in the bucket, bundles are also uploaded as `/bundle-archive/$commit-sha.json`. being able to have bundles available for each commit has several benefits, but this change is mostly driven by the idea to be able to run a dual-bundle qontract-server during app-interface PR checks by where one of the bundles represents the master branch at a verify specific commit.

the bundle-archive folder in the bucket has a lifecycle policy attached to cleanup old bundles after 90 days

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>