Crashloop pod alert only in onboarded services
Added dynamic rule type that uses a query to get all the onboarded
namespaces in a certain cluster.

In order to contain the scope of the tests, I also added skip
possibilities (that are maybe not fine grained enough but could be
helpful) to the new dynamic rule and also skip possbilities to metadata
tests to avoid touching the current tests.

Closes APPSRE-5949

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>