Move mem/cpu cause based alerts to quay slack
From AppSRE perspective, we're interested in symptom based alerts that
affect the whole service. The fact that a single pod is consuming a lot
of CPU or memory is not worrying. It is the service being affected as a
whole.

* CPU: The worst thing in this scenario is throttling.
* Memory: The worst case in this scenario is the pod getting OOMd and
  restarted. If pod is restarted many times, AppSRE will know.

The conditions are worrying if they are maintained in time but that
should be covered by other alerts that are service wide.

These alerts may be interesting for the service team, hence the severity
change. For AppSRE they are always no-ops that mainly generate noise in
the alerts channel.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>