Revert "patch - alert severity updated"
In the current state, this alert is going to trigger with any error that
Patchman returns at any given time. In this state it cannot be an alert
that gets into AppSRE channels as it is bound to be very flappy.

This alert should be rewritten with the recommendations in SPM-1621.

This reverts commit cae6e21336b517de0668aca848ff6a30f6065b85.