Reduce the frequency of DVO collection
We moved in the past from collecting every 24h to 1h, see
https://github.com/app-sre/qontract-reconcile/pull/2347.

But we're starting to see the DashdotDB volume getting out of
burst credit as the collection of the data is sometimes too demanding.
Given the nature of the alerts it finally drives and the long
resolution turnaround, I'd rather run it every 6 hours and give
the volume time to recover credits. The peak read IOPS are > 1k so
we're not going to get there via baseline performance, only by burst
performance.  The volume size is 40Gb and half of it is not used, so
I don't love the idea of increasing size for the sake of it.

part of APPSRE-6008.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>