Promote hive b87d3cb0a9879d567d573c8d47d550e3ba26a026
https://github.com/openshift/hive/compare/0c5e2fb547fea2697b2490a71ba2d03c90f16808...b87d3cb0a9879d567d573c8d47d550e3ba26a026

---
```
commit: 67a56a77107ce5556f970130fe24ab32a9f962d4
author: Eric Fried
Fix dnsendpoint/nameserverscraper cache

For existing DNSZones with copacetic name server entries in the root
hosted zone, restarting the controller was resulting in thrashing
because:

- the nameserverscraper would ignore subdomains not already in the
cache.
- the dnsendpoint controller would error if the scraper hadn't scraped
the subdomain yet.

This chicken/egg situation would eventually resolve itself because the
scraper would "seed" the cache entry for the subdomain on the first
pass. However, it would not resolve until the second pass of the scraper
two hours later.

This commit makes the dnsendpoint controller seed the cache when it
notices it's necessary. The scraper will then scrape for the subdomain
on its next pass *and* notify the dnsendpoint controller to reconcile
that DNSZone again.

We also needed to change the scraper to cache *all* subdomains for a
root domain, rather than trying to select only those for which this
controller is responsible (one root domain may be "shared" across
multiple hive shards). This is necessary because the scraper doesn't
know which subdomains those are beforehand. (It would theoretically be
nice to set up the controller to trigger the scraper when a
previously-unknown subdomain is encountered; but on controller restart
with a large number of existing DNSZones this would result in a massive
glut of requests to the cloud provider, which is exactly what the
nameserverscraper was designed to avoid.) This does increase the
memory footprint of the controller; but we calculated it to be on the
order of single-digit MiB with O(thousands) of "extraneous" subdomains,
which would be a fraction of a percent of the existing memory footprint;
and thus acceptable.

HIVE-1855

commit: 1bd777dd7e26f7fce93399b91cd7a7027641277b
author: Eric Fried
bundle-gen: Add --dummy-bundle COMMIT-ISH (for MCE)

Add support to hack/bundle-gen.py for generating bundle files locally,
from an arbitrary commit. The result is *not* a real OLM bundle; this is
explicitly for the case where hive is to be packaged into a larger
deliverable (specifically MCE). Of note, this option does not:
- Try to compute a real `version`. The `name` and `version` fields in
the CSV should be seen as placeholders, to be supplanted by the
consuming packaging.
- Build/push operator images.
- Talk to OperatorHub to a) calculate graph data (e.g. `replaces`), or
b) open PRs.

Example usage below. Note that the argument to `--dummy-bundle` can be
any commit-ish recognized by git, including a raw SHA.
```
[efried@efried hive]$ ./hack/bundle-gen.py --dummy-bundle origin/ocm-2.4
Cloning git@github.com:openshift/hive.git to /tmp/hive-repo-wwo7fkwx
Working in /tmp/hive-repo-wwo7fkwx
Checking out 74831db3a27301cc195ef4e78a43ebab36b10e9c
Skipping build of container quay.io/openshift-hive/hive:v0.0.3227-74831db
Writing bundle files to directory: /tmp/hive-operator-bundle-rq3_t5ua
Generating CSV for version: 0.0.3227-74831db
Wrote ClusterServiceVersion: /tmp/hive-operator-bundle-rq3_t5ua/0.0.3227-74831db/hive-operator.v0.0.3227-74831db.clusterserviceversion.yaml
Wrote bundle to /home/efried/go/src/github.com/openshift/hive/hive-operator-bundle-0.0.3227-74831db
[efried@efried hive]$ ll hive-operator-bundle-0.0.3227-74831db/
total 384
drwxrwxr-x.  2 efried efried  4096 Jul  7 10:16 ./
drwxrwxr-x. 18 efried efried  4096 Jul  7 10:16 ../
-rw-rw-r--.  1 efried efried  2062 Jul  7 10:16 hiveinternal.openshift.io_clustersyncleases.yaml
-rw-rw-r--.  1 efried efried 10938 Jul  7 10:16 hiveinternal.openshift.io_clustersyncs.yaml
-rw-rw-r--.  1 efried efried  7089 Jul  7 10:16 hiveinternal.openshift.io_fakeclusterinstalls.yaml
-rw-rw-r--.  1 efried efried  2783 Jul  7 10:16 hive.openshift.io_checkpoints.yaml
-rw-rw-r--.  1 efried efried  7542 Jul  7 10:16 hive.openshift.io_clusterclaims.yaml
-rw-rw-r--.  1 efried efried 57511 Jul  7 10:16 hive.openshift.io_clusterdeployments.yaml
-rw-rw-r--.  1 efried efried 14096 Jul  7 10:16 hive.openshift.io_clusterdeprovisions.yaml
-rw-rw-r--.  1 efried efried  2218 Jul  7 10:16 hive.openshift.io_clusterimagesets.yaml
-rw-rw-r--.  1 efried efried 30138 Jul  7 10:16 hive.openshift.io_clusterpools.yaml
-rw-rw-r--.  1 efried efried  7567 Jul  7 10:16 hive.openshift.io_clusterprovisions.yaml
-rw-rw-r--.  1 efried efried  5142 Jul  7 10:16 hive.openshift.io_clusterrelocates.yaml
-rw-rw-r--.  1 efried efried  4534 Jul  7 10:16 hive.openshift.io_clusterstates.yaml
-rw-rw-r--.  1 efried efried 11198 Jul  7 10:16 hive.openshift.io_dnszones.yaml
-rw-rw-r--.  1 efried efried 36393 Jul  7 10:16 hive.openshift.io_hiveconfigs.yaml
-rw-rw-r--.  1 efried efried  2577 Jul  7 10:16 hive.openshift.io_machinepoolnameleases.yaml
-rw-rw-r--.  1 efried efried 23401 Jul  7 10:16 hive.openshift.io_machinepools.yaml
-rw-rw-r--.  1 efried efried 36537 Jul  7 10:16 hive.openshift.io_selectorsyncidentityproviders.yaml
-rw-rw-r--.  1 efried efried  9783 Jul  7 10:16 hive.openshift.io_selectorsyncsets.yaml
-rw-rw-r--.  1 efried efried 34985 Jul  7 10:16 hive.openshift.io_syncidentityproviders.yaml
-rw-rw-r--.  1 efried efried  8176 Jul  7 10:16 hive.openshift.io_syncsets.yaml
-rw-rw-r--.  1 efried efried 34099 Jul  7 10:16 hive-operator.v0.0.3227-74831db.clusterserviceversion.yaml
```

HIVE-1954

commit: 60b4009be564212bd4e7ae7a010207fd422b774e
author: Eric Fried
Requeue when DNSZone cloud secret not yet copied

With managed DNS, the ClusterDeployment controller:
- Creates a DNSZone for the ClusterDeployment.
- On the first provision attempt, copies the cloud credentials secret
from the hive controllers namespace (based on
HiveConfig.Spec.ManagedDomains[].$Platform.CredentialsSecretRef) to the
namespace of the ClusterDeployment.

This can lead to a race condition whereby the DNSZone controller tries
to reconcile the DNSZone before that secret is in place, causing it to
error. However, in this case it was returning without requeueing, which
would result in the DNSZone never becoming ready, and the
ClusterDeployment would never provision.

Fix this by ensuring that code path (secret not found) returns an error,
ensuring it will requeue.

commit: 8a5f9f4a07e3c73f9e241d2dc106d4595559bf43
author: Eric Fried
Fix DNSZoneNotReady[TimedOut] hot loop

b7acaf2 / #1769 left a code path where the CD status condition
DNSZoneNotReady can thrash between two Reasons: DNSZoneNotReady and
DNSZoneNotReadyTimedOut. This pretty much defeats the intended purpose
of that effort, which was to make the latter a terminal state, whereupon
the controller should calm down and stop doing things.

This commit refactors the code to reduce the complexity, and resolves
the flawed logic.

HIVE-1881

```