Use IPv4 for OCM gateway SSO cluster
The OCM gateway implements the following URL redirectling it to
_sso.redhat.com:

  https://api.openshift.com/.well-known/jwks.json -> https://sso.redhat.com/auth/realms/redhat-external/protocol/openid-connect/certs

That has been deprecated long ago, but some services may still be using
it.

This started to fail recently, most probably because the
_sso.redhat.com_ service has been configured with IPv6 addresses that
can't be reached from the clusters where the gateway runs. Apparently
when Envoy sees those addresses it tries to use them even if the
environment where it runs doesn't support IPv6. To address that issue
this patch changes the Envoy configuration so that it always uses IPv4
for that communication.

This only affects the integration and stage environments. The production
environment will be updated in a separate patch.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>