design doc- describing changes in app-interface data
AppSRE needs to be capable of defining and testing new change types with low effort and use them to grant fine grained self-service permissions on `app-interface` data to tenants.

This design doc proposes describing types of changes as `app-interface` data and bind it to roles and datafiles/resources.

part of https://issues.redhat.com/browse/APPSRE-6095

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>