feature: sources monthly report
Management has requested us to pass them a monthly report with some
data from sources. This script should yield all the information we need.

The "schedule" key is commented because I would like to get the data for
this month too. The idea is to remove the comment in a follow-up commit.