no early exit when all integration should run on pr-checks
the select_all flag indicates that something changes where we can't detect the real involved integrations. this includes qr promotions when it is a good thing to not rely on desired changes to decide if an integration should be allowed to exit early. especially on qr promotions the pr check might reveal bugs in the integrations so it is crucial to run them.

maybe we find a better way to decide for such situations but this approach will lead to no false-positive early-exit decisions for now.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>