Retry on reset between stage OCM gateway and assisted installer
This patch changes the configuration of the OCM gateway of the stage
environment so that it will retry requests that fail because the
connection to the service (actually to the OpenShift router in front of
that service) is reset, as otherwise that results in a 503 error
resposne and an error message in the UI.

It is not clear why those connections are being reset. It may be because
the OpenShift router is restarted, or reloaded. Anyhow retrying in those
conditions doesn't hurt.

Related: https://issues.redhat.com/browse/MGMT-11711
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>